BitbankRestApiClientのユニットテストファイル名を修正
