+++ BitbankDotNet.Tests/PrivateApis/BitbankRestApiClientGetAssetsAsyncTest.cs
@@ -8,15 +8,15 @@ using System.Threading;
 using System.Threading.Tasks;
 using Xunit;
 
-namespace BitbankDotNet.Tests.PublicApis
+namespace BitbankDotNet.Tests.PrivateApis
 {
-    public class BitbankClientGetTickerAsyncTest
+    public class BitbankRestApiClientGetAssetsAsyncTest
     {
         const string Json =
-            "{\"success\":1,\"data\":{\"sell\":\"1.2\",\"buy\":\"1.2\",\"high\":\"1.2\",\"low\":\"1.2\",\"last\":\"1.2\",\"vol\":\"1.2\",\"timestamp\":1514862245678}}";
+            "{\"success\":1,\"data\":{\"assets\":[{\"asset\":\"jpy\",\"amount_precision\":3,\"onhand_amount\":\"1.2\",\"locked_amount\":\"1.2\",\"free_amount\":\"1.2\",\"withdrawal_fee\":{\"threshold\":\"1.2\",\"under\":\"1.2\",\"over\":\"1.2\"}},{\"asset\":\"jpy\",\"amount_precision\":3,\"onhand_amount\":\"1.2\",\"locked_amount\":\"1.2\",\"free_amount\":\"1.2\",\"withdrawal_fee\":{\"threshold\":\"1.2\",\"under\":\"1.2\",\"over\":\"1.2\"}}]}}";
 
         [Fact]
-        public void HTTPステータスが200かつSuccessが1_Tickerを返す()
+        public void HTTPステータスが200かつSuccessが1_Assetを返す()
         {
             var mockHttpHandler = new Mock<HttpMessageHandler>();
             mockHttpHandler.Protected()
@@ -24,7 +24,7 @@ namespace BitbankDotNet.Tests.PublicApis
                     ItExpr.IsAny<CancellationToken>())
                 .Callback<HttpRequestMessage, CancellationToken>((request, _) =>
                 {
-                    Assert.StartsWith("https://public.bitbank.cc/btc_jpy/", request.RequestUri.AbsoluteUri);
+					Assert.StartsWith("https://api.bitbank.cc/v1/user/", request.RequestUri.AbsoluteUri);
                 })
                 .ReturnsAsync(new HttpResponseMessage(HttpStatusCode.OK)
                 {
@@ -33,17 +33,19 @@ namespace BitbankDotNet.Tests.PublicApis
             
             using (var client = new HttpClient(mockHttpHandler.Object))
             {
-                var bitbank = new BitbankRestApiClient(client);
-                var result = bitbank.GetTickerAsync(default).GetAwaiter().GetResult();
+				var bitbank = new BitbankRestApiClient(client, " ", " ");
+                var result = bitbank.GetAssetsAsync().GetAwaiter().GetResult();
 
                 Assert.NotNull(result);
-                Assert.Equal(EntityHelper.GetTestValue<double>(), result.Buy);
-                Assert.Equal(EntityHelper.GetTestValue<double>(), result.High);
-                Assert.Equal(EntityHelper.GetTestValue<double>(), result.Last);
-                Assert.Equal(EntityHelper.GetTestValue<double>(), result.Low);
-                Assert.Equal(EntityHelper.GetTestValue<double>(), result.Sell);
-                Assert.Equal(EntityHelper.GetTestValue<DateTime>(), result.Timestamp);
-                Assert.Equal(EntityHelper.GetTestValue<double>(), result.Vol);
+                Assert.All(result, entity =>
+                {
+                    Assert.Equal(EntityHelper.GetTestValue<int>(), entity.AmountPrecision);
+                    Assert.Equal(EntityHelper.GetTestValue<double>(), entity.FreeAmount);
+                    Assert.Equal(EntityHelper.GetTestValue<double>(), entity.LockedAmount);
+                    Assert.Equal(EntityHelper.GetTestValue<AssetName>(), entity.Name);
+                    Assert.Equal(EntityHelper.GetTestValue<double>(), entity.OnhandAmount);
+                    Assert.Equal(EntityHelper.GetTestValue<WithdrawalFeeObject>(), entity.WithdrawalFee);
+                });
             }
         }
 
@@ -64,9 +66,9 @@ namespace BitbankDotNet.Tests.PublicApis
 
             using (var client = new HttpClient(mockHttpHandler.Object))
             {
-                var bitbank = new BitbankRestApiClient(client);
-                var exception = Assert.Throws<BitbankException>(() =>
-                    bitbank.GetTickerAsync(default).GetAwaiter().GetResult());
+				var bitbank = new BitbankRestApiClient(client, " ", " ");
+                var exception = Assert.Throws<BitbankApiException>(() =>
+                    bitbank.GetAssetsAsync().GetAwaiter().GetResult());
                 Assert.Equal(statusCode, exception.StatusCode);
                 Assert.Equal(apiErrorCode, exception.ApiErrorCode);
             }
@@ -90,9 +92,9 @@ namespace BitbankDotNet.Tests.PublicApis
 
             using (var client = new HttpClient(mockHttpHandler.Object))
             {
-                var bitbank = new BitbankRestApiClient(client, TimeSpan.FromMilliseconds(1));
-                var exception = Assert.Throws<BitbankException>(() =>
-                    bitbank.GetTickerAsync(default).GetAwaiter().GetResult());
+				var bitbank = new BitbankRestApiClient(client, " ", " ", TimeSpan.FromMilliseconds(1));
+                var exception = Assert.Throws<BitbankApiException>(() =>
+                    bitbank.GetAssetsAsync().GetAwaiter().GetResult());
                 Assert.IsType<TaskCanceledException>(exception.InnerException);
             }
         }
@@ -116,9 +118,9 @@ namespace BitbankDotNet.Tests.PublicApis
 
             using (var client = new HttpClient(mockHttpHandler.Object))
             {
-                var bitbank = new BitbankRestApiClient(client);
-                Assert.Throws<BitbankException>(() =>
-                    bitbank.GetTickerAsync(default).GetAwaiter().GetResult());
+				var bitbank = new BitbankRestApiClient(client, " ", " ");
+                Assert.Throws<BitbankApiException>(() =>
+                    bitbank.GetAssetsAsync().GetAwaiter().GetResult());
             }
         }
     }
