+++ BitbankDotNet.Tests/PrivateApis/BitbankRestApiClientGetOrdersAsyncTest.cs
@@ -8,15 +8,15 @@ using System.Threading;
 using System.Threading.Tasks;
 using Xunit;
 
-namespace BitbankDotNet.Tests.PublicApis
+namespace BitbankDotNet.Tests.PrivateApis
 {
-    public class BitbankClientGetCandlesticksAsyncTest
+    public class BitbankRestApiClientGetOrdersAsyncTest
     {
         const string Json =
-            "{\"success\":1,\"data\":{\"candlestick\":[{\"type\":\"1min\",\"ohlcv\":[[\"1.2\",\"1.2\",\"1.2\",\"1.2\",\"1.2\",1514862245678],[\"1.2\",\"1.2\",\"1.2\",\"1.2\",\"1.2\",1514862245678]]},{\"type\":\"1min\",\"ohlcv\":[[\"1.2\",\"1.2\",\"1.2\",\"1.2\",\"1.2\",1514862245678],[\"1.2\",\"1.2\",\"1.2\",\"1.2\",\"1.2\",1514862245678]]}]}}";
+            "{\"success\":1,\"data\":{\"orders\":[{\"order_id\":4,\"pair\":\"btc_jpy\",\"side\":\"buy\",\"type\":\"limit\",\"start_amount\":\"1.2\",\"remaining_amount\":\"1.2\",\"executed_amount\":\"1.2\",\"price\":\"1.2\",\"average_price\":\"1.2\",\"ordered_at\":1514862245678,\"status\":\"UNFILLED\"},{\"order_id\":4,\"pair\":\"btc_jpy\",\"side\":\"buy\",\"type\":\"limit\",\"start_amount\":\"1.2\",\"remaining_amount\":\"1.2\",\"executed_amount\":\"1.2\",\"price\":\"1.2\",\"average_price\":\"1.2\",\"ordered_at\":1514862245678,\"status\":\"UNFILLED\"}]}}";
 
         [Fact]
-        public void HTTPステータスが200かつSuccessが1_Ohlcvを返す()
+        public void HTTPステータスが200かつSuccessが1_Orderを返す()
         {
             var mockHttpHandler = new Mock<HttpMessageHandler>();
             mockHttpHandler.Protected()
@@ -24,7 +24,7 @@ namespace BitbankDotNet.Tests.PublicApis
                     ItExpr.IsAny<CancellationToken>())
                 .Callback<HttpRequestMessage, CancellationToken>((request, _) =>
                 {
-                    Assert.StartsWith("https://public.bitbank.cc/btc_jpy/", request.RequestUri.AbsoluteUri);
+					Assert.StartsWith("https://api.bitbank.cc/v1/user/", request.RequestUri.AbsoluteUri);
                 })
                 .ReturnsAsync(new HttpResponseMessage(HttpStatusCode.OK)
                 {
@@ -33,18 +33,23 @@ namespace BitbankDotNet.Tests.PublicApis
             
             using (var client = new HttpClient(mockHttpHandler.Object))
             {
-                var bitbank = new BitbankRestApiClient(client);
-                var result = bitbank.GetCandlesticksAsync(default, default, default, default, default).GetAwaiter().GetResult();
+				var bitbank = new BitbankRestApiClient(client, " ", " ");
+                var result = bitbank.GetOrdersAsync(default, default).GetAwaiter().GetResult();
 
                 Assert.NotNull(result);
                 Assert.All(result, entity =>
                 {
-                    Assert.Equal(EntityHelper.GetTestValue<double>(), entity.Close);
-                    Assert.Equal(EntityHelper.GetTestValue<DateTime>(), entity.Date);
-                    Assert.Equal(EntityHelper.GetTestValue<double>(), entity.High);
-                    Assert.Equal(EntityHelper.GetTestValue<double>(), entity.Low);
-                    Assert.Equal(EntityHelper.GetTestValue<double>(), entity.Open);
-                    Assert.Equal(EntityHelper.GetTestValue<double>(), entity.Volume);
+                    Assert.Equal(EntityHelper.GetTestValue<double>(), entity.AveragePrice);
+                    Assert.Equal(EntityHelper.GetTestValue<double>(), entity.ExecutedAmount);
+                    Assert.Equal(EntityHelper.GetTestValue<DateTime>(), entity.OrderedAt);
+                    Assert.Equal(EntityHelper.GetTestValue<long>(), entity.OrderId);
+                    Assert.Equal(EntityHelper.GetTestValue<CurrencyPair>(), entity.Pair);
+                    Assert.Equal(EntityHelper.GetTestValue<double>(), entity.Price);
+                    Assert.Equal(EntityHelper.GetTestValue<double>(), entity.RemainingAmount);
+                    Assert.Equal(EntityHelper.GetTestValue<OrderSide>(), entity.Side);
+                    Assert.Equal(EntityHelper.GetTestValue<double>(), entity.StartAmount);
+                    Assert.Equal(EntityHelper.GetTestValue<OrderStatus>(), entity.Status);
+                    Assert.Equal(EntityHelper.GetTestValue<OrderType>(), entity.Type);
                 });
             }
         }
@@ -66,9 +71,9 @@ namespace BitbankDotNet.Tests.PublicApis
 
             using (var client = new HttpClient(mockHttpHandler.Object))
             {
-                var bitbank = new BitbankRestApiClient(client);
-                var exception = Assert.Throws<BitbankException>(() =>
-                    bitbank.GetCandlesticksAsync(default, default, default, default, default).GetAwaiter().GetResult());
+				var bitbank = new BitbankRestApiClient(client, " ", " ");
+                var exception = Assert.Throws<BitbankApiException>(() =>
+                    bitbank.GetOrdersAsync(default, default).GetAwaiter().GetResult());
                 Assert.Equal(statusCode, exception.StatusCode);
                 Assert.Equal(apiErrorCode, exception.ApiErrorCode);
             }
@@ -92,9 +97,9 @@ namespace BitbankDotNet.Tests.PublicApis
 
             using (var client = new HttpClient(mockHttpHandler.Object))
             {
-                var bitbank = new BitbankRestApiClient(client, TimeSpan.FromMilliseconds(1));
-                var exception = Assert.Throws<BitbankException>(() =>
-                    bitbank.GetCandlesticksAsync(default, default, default, default, default).GetAwaiter().GetResult());
+				var bitbank = new BitbankRestApiClient(client, " ", " ", TimeSpan.FromMilliseconds(1));
+                var exception = Assert.Throws<BitbankApiException>(() =>
+                    bitbank.GetOrdersAsync(default, default).GetAwaiter().GetResult());
                 Assert.IsType<TaskCanceledException>(exception.InnerException);
             }
         }
@@ -118,9 +123,9 @@ namespace BitbankDotNet.Tests.PublicApis
 
             using (var client = new HttpClient(mockHttpHandler.Object))
             {
-                var bitbank = new BitbankRestApiClient(client);
-                Assert.Throws<BitbankException>(() =>
-                    bitbank.GetCandlesticksAsync(default, default, default, default, default).GetAwaiter().GetResult());
+				var bitbank = new BitbankRestApiClient(client, " ", " ");
+                Assert.Throws<BitbankApiException>(() =>
+                    bitbank.GetOrdersAsync(default, default).GetAwaiter().GetResult());
             }
         }
     }
