+++ BitbankDotNet.Tests/PublicApis/BitbankRestApiClientGetCandlesticksAsyncTest.cs
@@ -8,15 +8,15 @@ using System.Threading;
 using System.Threading.Tasks;
 using Xunit;
 
-namespace BitbankDotNet.Tests.PrivateApis
+namespace BitbankDotNet.Tests.PublicApis
 {
-    public class BitbankClientGetOrderAsyncTest
+    public class BitbankRestApiClientGetCandlesticksAsyncTest
     {
         const string Json =
-            "{\"success\":1,\"data\":{\"order_id\":4,\"pair\":\"btc_jpy\",\"side\":\"buy\",\"type\":\"limit\",\"start_amount\":\"1.2\",\"remaining_amount\":\"1.2\",\"executed_amount\":\"1.2\",\"price\":\"1.2\",\"average_price\":\"1.2\",\"ordered_at\":1514862245678,\"status\":\"UNFILLED\"}}";
+            "{\"success\":1,\"data\":{\"candlestick\":[{\"type\":\"1min\",\"ohlcv\":[[\"1.2\",\"1.2\",\"1.2\",\"1.2\",\"1.2\",1514862245678],[\"1.2\",\"1.2\",\"1.2\",\"1.2\",\"1.2\",1514862245678]]},{\"type\":\"1min\",\"ohlcv\":[[\"1.2\",\"1.2\",\"1.2\",\"1.2\",\"1.2\",1514862245678],[\"1.2\",\"1.2\",\"1.2\",\"1.2\",\"1.2\",1514862245678]]}]}}";
 
         [Fact]
-        public void HTTPステータスが200かつSuccessが1_Orderを返す()
+        public void HTTPステータスが200かつSuccessが1_Ohlcvを返す()
         {
             var mockHttpHandler = new Mock<HttpMessageHandler>();
             mockHttpHandler.Protected()
@@ -24,7 +24,7 @@ namespace BitbankDotNet.Tests.PrivateApis
                     ItExpr.IsAny<CancellationToken>())
                 .Callback<HttpRequestMessage, CancellationToken>((request, _) =>
                 {
-					Assert.StartsWith("https://api.bitbank.cc/v1/user/", request.RequestUri.AbsoluteUri);
+                    Assert.StartsWith("https://public.bitbank.cc/btc_jpy/", request.RequestUri.AbsoluteUri);
                 })
                 .ReturnsAsync(new HttpResponseMessage(HttpStatusCode.OK)
                 {
@@ -33,21 +33,19 @@ namespace BitbankDotNet.Tests.PrivateApis
             
             using (var client = new HttpClient(mockHttpHandler.Object))
             {
-				var bitbank = new BitbankRestApiClient(client, " ", " ");
-                var result = bitbank.GetOrderAsync(default, default).GetAwaiter().GetResult();
+                var bitbank = new BitbankRestApiClient(client);
+                var result = bitbank.GetCandlesticksAsync(default, default, default, default, default).GetAwaiter().GetResult();
 
                 Assert.NotNull(result);
-                Assert.Equal(EntityHelper.GetTestValue<double>(), result.AveragePrice);
-                Assert.Equal(EntityHelper.GetTestValue<double>(), result.ExecutedAmount);
-                Assert.Equal(EntityHelper.GetTestValue<DateTime>(), result.OrderedAt);
-                Assert.Equal(EntityHelper.GetTestValue<long>(), result.OrderId);
-                Assert.Equal(EntityHelper.GetTestValue<CurrencyPair>(), result.Pair);
-                Assert.Equal(EntityHelper.GetTestValue<double>(), result.Price);
-                Assert.Equal(EntityHelper.GetTestValue<double>(), result.RemainingAmount);
-                Assert.Equal(EntityHelper.GetTestValue<OrderSide>(), result.Side);
-                Assert.Equal(EntityHelper.GetTestValue<double>(), result.StartAmount);
-                Assert.Equal(EntityHelper.GetTestValue<OrderStatus>(), result.Status);
-                Assert.Equal(EntityHelper.GetTestValue<OrderType>(), result.Type);
+                Assert.All(result, entity =>
+                {
+                    Assert.Equal(EntityHelper.GetTestValue<double>(), entity.Close);
+                    Assert.Equal(EntityHelper.GetTestValue<DateTime>(), entity.Date);
+                    Assert.Equal(EntityHelper.GetTestValue<double>(), entity.High);
+                    Assert.Equal(EntityHelper.GetTestValue<double>(), entity.Low);
+                    Assert.Equal(EntityHelper.GetTestValue<double>(), entity.Open);
+                    Assert.Equal(EntityHelper.GetTestValue<double>(), entity.Volume);
+                });
             }
         }
 
@@ -68,9 +66,9 @@ namespace BitbankDotNet.Tests.PrivateApis
 
             using (var client = new HttpClient(mockHttpHandler.Object))
             {
-				var bitbank = new BitbankRestApiClient(client, " ", " ");
-                var exception = Assert.Throws<BitbankException>(() =>
-                    bitbank.GetOrderAsync(default, default).GetAwaiter().GetResult());
+                var bitbank = new BitbankRestApiClient(client);
+                var exception = Assert.Throws<BitbankApiException>(() =>
+                    bitbank.GetCandlesticksAsync(default, default, default, default, default).GetAwaiter().GetResult());
                 Assert.Equal(statusCode, exception.StatusCode);
                 Assert.Equal(apiErrorCode, exception.ApiErrorCode);
             }
@@ -94,9 +92,9 @@ namespace BitbankDotNet.Tests.PrivateApis
 
             using (var client = new HttpClient(mockHttpHandler.Object))
             {
-				var bitbank = new BitbankRestApiClient(client, " ", " ", TimeSpan.FromMilliseconds(1));
-                var exception = Assert.Throws<BitbankException>(() =>
-                    bitbank.GetOrderAsync(default, default).GetAwaiter().GetResult());
+                var bitbank = new BitbankRestApiClient(client, TimeSpan.FromMilliseconds(1));
+                var exception = Assert.Throws<BitbankApiException>(() =>
+                    bitbank.GetCandlesticksAsync(default, default, default, default, default).GetAwaiter().GetResult());
                 Assert.IsType<TaskCanceledException>(exception.InnerException);
             }
         }
@@ -120,9 +118,9 @@ namespace BitbankDotNet.Tests.PrivateApis
 
             using (var client = new HttpClient(mockHttpHandler.Object))
             {
-				var bitbank = new BitbankRestApiClient(client, " ", " ");
-                Assert.Throws<BitbankException>(() =>
-                    bitbank.GetOrderAsync(default, default).GetAwaiter().GetResult());
+                var bitbank = new BitbankRestApiClient(client);
+                Assert.Throws<BitbankApiException>(() =>
+                    bitbank.GetCandlesticksAsync(default, default, default, default, default).GetAwaiter().GetResult());
             }
         }
     }
